# Fix coalescing of pcmpestri/pcmpestrm instructions