I used [cargo-asm](https://github.com/gnzlbg/cargo-asm) to check the generated assembly for the following program:

```rust
#[macro_use] extern crate jetscii;

fn main() {
    let part_number = "                                                         86-J52:rev1";
    let first = ascii_chars!('-', ':').find(part_number);
    assert_eq!(first, Some(2));
}
```

After making `PackedCompare<T>::cmpestri` and `PackedCompare<T>::cmpestrm` inline never so that viewing the generated assembly would be easier, I found that the two pcmpestri/pcmpestrm instructions (one to check if a match was found, and the other to handle the found index) were not coalesced.

`cargo asm '<jetscii::simd::PackedCompare<T>>::cmpestri' --rust` generated:

```asm
<jetscii::simd::PackedCompare<T>>::cmpestri (/Users/dtrebbien/Documents/Projects/jetscii/src/simd.rs:148):
 push    rbp
 mov     rbp, rsp
 push    r15
 push    r14
 push    rbx
 sub     rsp, 56
 mov     r14d, edx
 mov     rbx, rdi
 movups  xmm0, xmmword, ptr, [rsi]
 movaps  xmmword, ptr, [rbp, -, 80], xmm0
 lea     rdi, [rbp, -, 64]
 mov     rsi, rbx
 call    <&'b jetscii::simd::Bytes as jetscii::simd::PackedCompareControl>::needle
 movaps  xmm0, xmmword, ptr, [rbp, -, 64]
 movaps  xmmword, ptr, [rbp, -, 48], xmm0
 mov     rdi, rbx
 call    <&'b jetscii::simd::Bytes as jetscii::simd::PackedCompareControl>::needle_len
 xor     r15d, r15d
 mov     edx, r14d
 movdqa  xmm0, xmmword, ptr, [rbp, -, 48]
 pcmpestri xmm0, xmmword, ptr, [rbp, -, 80], 0
 setb    r15b
 lea     rdi, [rbp, -, 64]
 mov     rsi, rbx
 call    <&'b jetscii::simd::Bytes as jetscii::simd::PackedCompareControl>::needle
 movaps  xmm0, xmmword, ptr, [rbp, -, 64]
 movaps  xmmword, ptr, [rbp, -, 48], xmm0
 mov     rdi, rbx
 call    <&'b jetscii::simd::Bytes as jetscii::simd::PackedCompareControl>::needle_len
 mov     edx, r14d
 movdqa  xmm0, xmmword, ptr, [rbp, -, 48]
 pcmpestri xmm0, xmmword, ptr, [rbp, -, 80], 0
 movsxd  rdx, ecx
 mov     rax, r15
 add     rsp, 56
 pop     rbx
 pop     r14
 pop     r15
 pop     rbp
 ret
```

`cargo asm '<jetscii::simd::PackedCompare<T>>::cmpestrm' --rust` generated:

```asm
<jetscii::simd::PackedCompare<T>>::cmpestrm (/Users/dtrebbien/Documents/Projects/jetscii/src/simd.rs:101):
 push    rbp
 mov     rbp, rsp
 push    r15
 push    r14
 push    rbx
 sub     rsp, 72
 mov     r14, rdx
 mov     rbx, rdi
 movups  xmm0, xmmword, ptr, [rsi]
 movaps  xmmword, ptr, [rbp, -, 96], xmm0
 lea     rdi, [rbp, -, 48]
 mov     rsi, rbx
 call    <&'b jetscii::simd::Bytes as jetscii::simd::PackedCompareControl>::needle
 movaps  xmm0, xmmword, ptr, [rbp, -, 48]
 movaps  xmmword, ptr, [rbp, -, 80], xmm0
 mov     rdi, rbx
 call    <&'b jetscii::simd::Bytes as jetscii::simd::PackedCompareControl>::needle_len
 mov     r15d, eax
 lea     rdi, [rbp, -, 48]
 mov     rsi, rbx
 call    <&'b jetscii::simd::Bytes as jetscii::simd::PackedCompareControl>::needle
 movaps  xmm0, xmmword, ptr, [rbp, -, 48]
 movaps  xmmword, ptr, [rbp, -, 64], xmm0
 mov     rdi, rbx
 call    <&'b jetscii::simd::Bytes as jetscii::simd::PackedCompareControl>::needle_len
 movdqa  xmm0, xmmword, ptr, [rbp, -, 96]
 mov     esi, eax
 mov     edx, 16
 mov     eax, r15d
 movdqa  xmm1, xmmword, ptr, [rbp, -, 80]
 pcmpestri xmm1, xmm0, 0
 jae     LBB6_1
 mov     edx, 16
 mov     eax, esi
 movdqa  xmm1, xmmword, ptr, [rbp, -, 64]
 pcmpestrm xmm1, xmm0, 0
 movq    rax, xmm0
 mov     ecx, r14d
 shr     rax, cl
 test    rax, rax
 je      LBB6_1
 je      LBB6_4
 bsf     rdx, rax
 jmp     LBB6_6
LBB6_1:
 xor     eax, eax
 jmp     LBB6_7
LBB6_4:
 mov     edx, 64
LBB6_6:
 mov     eax, 1
LBB6_7:
 add     rsp, 72
 pop     rbx
 pop     r14
 pop     r15
 pop     rbp
 ret
```

With this PR applied, one additional pcmpestri instruction is eliminated:

```asm
<jetscii::simd::PackedCompare<T>>::cmpestri (/Users/dtrebbien/Documents/Projects/jetscii/src/simd.rs:139):
 push    rbp
 mov     rbp, rsp
 push    r14
 push    rbx
 sub     rsp, 48
 mov     r14d, edx
 mov     rbx, rdi
 movups  xmm0, xmmword, ptr, [rsi]
 movaps  xmmword, ptr, [rbp, -, 48], xmm0
 lea     rdi, [rbp, -, 64]
 mov     rsi, rbx
 call    <&'b jetscii::simd::Bytes as jetscii::simd::PackedCompareControl>::needle
 movaps  xmm0, xmmword, ptr, [rbp, -, 64]
 movaps  xmmword, ptr, [rbp, -, 32], xmm0
 mov     rdi, rbx
 call    <&'b jetscii::simd::Bytes as jetscii::simd::PackedCompareControl>::needle_len
 mov     edx, r14d
 movdqa  xmm0, xmmword, ptr, [rbp, -, 32]
 pcmpestri xmm0, xmmword, ptr, [rbp, -, 48], 0
 xor     eax, eax
 cmp     ecx, 16
 setl    al
 movsxd  rdx, ecx
 add     rsp, 48
 pop     rbx
 pop     r14
 pop     rbp
 ret
```

```asm
<jetscii::simd::PackedCompare<T>>::cmpestrm (/Users/dtrebbien/Documents/Projects/jetscii/src/simd.rs:100):
 push    rbp
 mov     rbp, rsp
 push    r14
 push    rbx
 sub     rsp, 48
 mov     r14, rdx
 mov     rbx, rdi
 movups  xmm0, xmmword, ptr, [rsi]
 movaps  xmmword, ptr, [rbp, -, 48], xmm0
 lea     rdi, [rbp, -, 64]
 mov     rsi, rbx
 call    <&'b jetscii::simd::Bytes as jetscii::simd::PackedCompareControl>::needle
 movaps  xmm0, xmmword, ptr, [rbp, -, 64]
 movaps  xmmword, ptr, [rbp, -, 32], xmm0
 mov     rdi, rbx
 call    <&'b jetscii::simd::Bytes as jetscii::simd::PackedCompareControl>::needle_len
 mov     edx, 16
 movdqa  xmm0, xmmword, ptr, [rbp, -, 32]
 pcmpestrm xmm0, xmmword, ptr, [rbp, -, 48], 0
 movd    eax, xmm0
 test    ax, ax
 je      LBB6_1
 bsf     cx, ax
 movzx   ecx, cx
 cmp     ecx, 15
 jbe     LBB6_5
 jmp     LBB6_4
LBB6_1:
 mov     cx, 16
 movzx   ecx, cx
 cmp     ecx, 15
 ja      LBB6_4
LBB6_5:
 movzx   eax, ax
 and     r14b, 15
 mov     ecx, r14d
 shr     eax, cl
 test    ax, ax
 je      LBB6_4
 je      LBB6_7
 bsf     ax, ax
 jmp     LBB6_9
LBB6_4:
 xor     eax, eax
 jmp     LBB6_10
LBB6_7:
 mov     ax, 16
LBB6_9:
 movzx   edx, ax
 mov     eax, 1
LBB6_10:
 add     rsp, 48
 pop     rbx
 pop     r14
 pop     rbp
 ret
```

I ran the benchmarks using rustc 1.30.0-nightly (73c78734b 2018-08-05) on macOS 10.13.6 with a 2.3 GHz Intel Core i7, and the Jetscii benchmarks improve slightly:

Before:

**test bench::space_ascii_chars                    ... bench:     793,830 ns/iter (+/- 38,661) = 6604 MB/s**  
test bench::space_stdlib_find_char               ... bench:     537,295 ns/iter (+/- 119,061) = 9757 MB/s  
test bench::space_stdlib_find_char_set           ... bench:   7,074,271 ns/iter (+/- 163,336) = 741 MB/s  
test bench::space_stdlib_find_closure            ... bench:   7,105,007 ns/iter (+/- 354,860) = 737 MB/s  
test bench::space_stdlib_find_string             ... bench:   3,190,072 ns/iter (+/- 108,691) = 1643 MB/s  
test bench::space_stdlib_iterator_position       ... bench:   3,162,603 ns/iter (+/- 75,241) = 1657 MB/s  
test bench::substring_stdlib_find                ... bench:     989,340 ns/iter (+/- 58,210) = 5299 MB/s  
**test bench::substring_with_created_searcher      ... bench:     799,166 ns/iter (+/- 38,097) = 6560 MB/s  
test bench::xml_delim_3_ascii_chars              ... bench:     795,045 ns/iter (+/- 38,551) = 6594 MB/s**  
test bench::xml_delim_3_stdlib_find_char_closure ... bench:   7,080,469 ns/iter (+/- 173,507) = 740 MB/s  
test bench::xml_delim_3_stdlib_find_char_set     ... bench:   7,351,264 ns/iter (+/- 264,844) = 713 MB/s  
test bench::xml_delim_3_stdlib_iterator_position ... bench:   4,337,162 ns/iter (+/- 139,146) = 1208 MB/s  
**test bench::xml_delim_5_ascii_chars              ... bench:     804,849 ns/iter (+/- 62,880) = 6514 MB/s**  
test bench::xml_delim_5_stdlib_find_char_closure ... bench:   7,058,137 ns/iter (+/- 158,827) = 742 MB/s  
test bench::xml_delim_5_stdlib_find_char_set     ... bench:  12,977,299 ns/iter (+/- 1,896,773) = 404 MB/s  
test bench::xml_delim_5_stdlib_iterator_position ... bench:   8,827,361 ns/iter (+/- 297,306) = 593 MB/s

After:

**test bench::space_ascii_chars                    ... bench:     777,229 ns/iter (+/- 35,709) = 6745 MB/s**  
test bench::space_stdlib_find_char               ... bench:     548,953 ns/iter (+/- 123,592) = 9550 MB/s  
test bench::space_stdlib_find_char_set           ... bench:   7,080,518 ns/iter (+/- 226,482) = 740 MB/s  
test bench::space_stdlib_find_closure            ... bench:   7,049,567 ns/iter (+/- 181,421) = 743 MB/s  
test bench::space_stdlib_find_string             ... bench:   3,174,222 ns/iter (+/- 103,375) = 1651 MB/s  
test bench::space_stdlib_iterator_position       ... bench:   3,154,285 ns/iter (+/- 85,194) = 1662 MB/s  
test bench::substring_stdlib_find                ... bench:     973,902 ns/iter (+/- 67,721) = 5383 MB/s  
**test bench::substring_with_created_searcher      ... bench:     792,693 ns/iter (+/- 38,922) = 6614 MB/s  
test bench::xml_delim_3_ascii_chars              ... bench:     794,039 ns/iter (+/- 31,857) = 6602 MB/s**  
test bench::xml_delim_3_stdlib_find_char_closure ... bench:   7,044,328 ns/iter (+/- 176,813) = 744 MB/s  
test bench::xml_delim_3_stdlib_find_char_set     ... bench:   7,125,839 ns/iter (+/- 345,773) = 735 MB/s  
test bench::xml_delim_3_stdlib_iterator_position ... bench:   4,360,698 ns/iter (+/- 359,530) = 1202 MB/s  
**test bench::xml_delim_5_ascii_chars              ... bench:     794,298 ns/iter (+/- 39,496) = 6600 MB/s**  
test bench::xml_delim_5_stdlib_find_char_closure ... bench:   7,054,105 ns/iter (+/- 157,787) = 743 MB/s  
test bench::xml_delim_5_stdlib_find_char_set     ... bench:  13,108,595 ns/iter (+/- 885,809) = 399 MB/s  
test bench::xml_delim_5_stdlib_iterator_position ... bench:   8,662,344 ns/iter (+/- 192,292) = 605 MB/s