@shepmaster 2018-08-06
Thanks! This certainly seems better, so I'm definitely going to accept this. However, I swear I checked this (comments to the opposite notwithstanding...). I [definitely had some discussion about it](https://github.com/rust-lang-nursery/stdsimd/issues/360#issuecomment-371628375).

When executing the PCMPxSTRx instruction, it already sets various CPU flags indicating extra outputs. Ideally, these flags are what the `if` statement would be using, which should avoid even doing a second comparison.

Any thoughts on such a path?