@dtrebbien 2018-08-06
As `PackedCompare<T>::cmpestri` seems pretty much the same to me in the "before" vs. "after" assembly (except for one fewer PCMPESTRI instruction), I think that it's best to focus on `PackedCompare<T>::cmpestrm`, because that's where I see an unfortunate side effect of this change.

As I understand it, the PCMPESTRM instruction sets the CFlag to 0 if there wasn't a match, and the _mm_cmpestrc() intrinsic represents the value of the CFlag after a PCMPESTRM instruction. We can see this in the "before" assembly of `PackedCompare<T>::cmpestrm`; this code:

```rust
        let found = _mm_cmpestrc(
            self.0.needle(),
            self.0.needle_len(),
            haystack,
            BYTES_PER_OPERATION as i32,
            T::CONTROL_BYTE,
        );
        //...

        if found != 0 {
            //...
        } else {
            None
        }
```

.. was compiled to:

```asm
 pcmpestri xmm1, xmm0, 0 ; this computes `found` by way of the CFlag
 jae     LBB6_1
 ;...
LBB6_1:
 ;...
```

The JAE instruction causes control flow to jump over the second PCMPESTRM instruction (which computes `mask`) in the event that CFlag is 0.

In the "after" assembly, the PCMPESTRI & JAE instructions became:

```asm
 pcmpestrm xmm0, xmmword, ptr, [rbp, -, 48], 0
 movd    eax, xmm0
 test    ax, ax
 je      LBB6_1
```

In the event that there wasn't a match, the MOVD instruction is somewhat wasteful because it copies data from an XMM register to EAX, as opposed to just checking the CFlag. If there were a way to access the CFlag *and* output XMM register of the PCMPESTRM instruction through an instrinsic, that would be ideal.

In trying to find out how to access the EFLAGS register in Rust stdsimd, I found rust-lang-nursery/stdsimd#485, which concerns the now-deprecated \__readeflags() and \__writeeflags() intrinsics. @rkruppe makes the cogent point that it's possible for LLVM to move code before the PUSHFQ & POPQ instructions generated by a \__readeflags() intrinsic; therefore, LLVM might move EFLAGS-clobbering code between the PCMPESTRM instruction generated by an \_mm_cmpestrm() instrinsic and the PUSHFQ & POPQ instructions generated by a \__readeflags() intrinsic.

However, even if \__readeflags() were safe to use, I think that the PUSHFQ & POPQ instructions are also less than ideal.

I think that the only way to get a PCMPESTRM instruction immediately followed by a JAE instruction is to use inline assembly, but this is not allowed on Rust stable.

Fortunately, it looks like `PackedCompare<T>::cmpestrm` is called at most once per find operation, and only when the haystack is not aligned.