@dtrebbien 2018-08-06
By the way, with respect to the re-licensing issue (#12), I hereby license this contribution under the dual MIT/Apache-2.0 license, allowing licensees to choose either at their option.