@shepmaster 2018-08-06
> By the way, with respect to the re-licensing issue (#12),

Haha, I had totally forgotten about that. Thank you for reminding me 😅