@shepmaster 2018-08-06
Thank you for that wonderfully deep exploration!

> get a PCMPESTRM instruction immediately followed by a JAE instruction is to use inline assembly

The only reason I know about this is because that's what the code *used* to do, but your benchmarks show that this PR is definitely faster than what we have now, so forward!