@shepmaster 2018-08-06
I wonder if it is worth raising a stdsimd issue that back-to-back calls of the same instruction weren't coalesced...