@shepmaster 2018-08-06
I'm going to ping the people on the relicense issue to try and update that for the next release, but if I don't hear from them in a few days or so, I'll release a new version with the current license. Please feel free to poke me if I forget (a common occurrence)!