@dtrebbien 2018-08-07
Thank you, Jake.

> I wonder if it is worth raising a stdsimd issue that back-to-back calls of the same instruction weren't coalesced...

I think so. I'll try to put together a minimal test case with the output from rustc stable and nightly, as well as what Clang and maybe GCC do in similar code.